+++ xla/service/gpu/runtime/command_buffer_cmd_emitter.cc
@@ -29,7 +29,6 @@ limitations under the License.
 #include "xla/service/gpu/runtime/copy_thunk.h"
 #include "xla/service/gpu/runtime/cudnn_thunk.h"
 #include "xla/service/gpu/runtime/custom_call_thunk.h"
-#include "xla/service/gpu/runtime/fused_mha_thunk.h"
 #include "xla/service/gpu/runtime/gemm_thunk.h"
 #include "xla/service/gpu/runtime/gpublas_lt_matmul_thunk.h"
 #include "xla/service/gpu/runtime/kernel_thunk.h"
@@ -141,27 +140,6 @@ static absl::StatusOr<Command> Convert(const CublasLtMatmulThunk& thunk) {
       thunk.aux_buffer(), thunk.a_scale_buffer(), thunk.b_scale_buffer(),
       thunk.c_scale_buffer(), thunk.d_scale_buffer(), thunk.d_amax_buffer(),
       thunk.workspace().value());
-}
-
-static absl::StatusOr<Command> Convert(const FusedMHAThunk& thunk) {
-  return std::make_unique<FusedMHACmd>(
-      thunk.execution_stream_id(), thunk.config(), thunk.lhs_bmm1_buffer(),
-      thunk.rhs_bmm1_buffer(), thunk.rhs_bmm2_buffer(), thunk.output_buffer(),
-      thunk.scratch_buffer(), BufferAllocation::Slice(), thunk.bias_buffer(),
-      thunk.activation_buffer(), thunk.seqlen_q_buffer(),
-      thunk.seqlen_k_buffer());
-}
-
-static absl::StatusOr<Command> Convert(const FusedMHABackwardThunk& thunk) {
-  return std::make_unique<FusedMHABackwardCmd>(
-      thunk.execution_stream_id(), thunk.config(),
-      thunk.bmm1_grad_gemm1_rhs_buffer(), thunk.bmm1_grad_gemm2_rhs_buffer(),
-      thunk.bmm2_grad_gemm1_lhs_buffer(), thunk.bmm2_grad_gemm2_rhs_buffer(),
-      thunk.d_output_buffer(), thunk.scratch_buffer(),
-      thunk.d_bmm1_lhs_buffer(), thunk.d_bmm1_rhs_buffer(),
-      thunk.d_bmm2_rhs_buffer(), thunk.d_s_buffer(), thunk.d_bias_buffer(),
-      thunk.fwd_output_buffer(), thunk.bias_buffer(), thunk.seqlen_q_buffer(),
-      thunk.seqlen_k_buffer());
 }
 
 static absl::StatusOr<Command> Convert(
@@ -276,10 +254,6 @@ static absl::Status AppendCommands(
       return append(Convert<CustomCallThunk>(thunk));
     case Thunk::Kind::kCustomKernel:
       return append(Convert<CustomKernelThunk>(thunk));
-    case Thunk::Kind::kFusedMHA:
-      return append(Convert<FusedMHAThunk>(thunk));
-    case Thunk::Kind::kFusedMHABackward:
-      return append(Convert<FusedMHABackwardThunk>(thunk));
     case Thunk::Kind::kKernel:
       return append(Convert<KernelThunk>(thunk));
     case Thunk::Kind::kGemm:
