+++ xla/service/gpu/transforms/cudnn_custom_call_compiler.h
@@ -13,13 +13,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-#ifndef XLA_SERVICE_GPU_TRANSFORMS_CUDNN_WORKSPACE_REWRITER_H_
-#define XLA_SERVICE_GPU_TRANSFORMS_CUDNN_WORKSPACE_REWRITER_H_
+#ifndef XLA_SERVICE_GPU_TRANSFORMS_CUDNN_CUSTOM_CALL_COMPILER_H_
+#define XLA_SERVICE_GPU_TRANSFORMS_CUDNN_CUSTOM_CALL_COMPILER_H_
 
 #include "absl/container/flat_hash_set.h"
 #include "absl/status/statusor.h"
 #include "absl/strings/string_view.h"
 #include "xla/hlo/ir/hlo_module.h"
+#include "xla/service/gpu/ir_emission_utils.h"
 #include "xla/service/hlo_pass_interface.h"
 #include "xla/stream_executor/dnn.h"
 #include "xla/stream_executor/stream_executor.h"
@@ -27,14 +28,18 @@ limitations under the License.
 namespace xla {
 namespace gpu {
 
-// Rewrite cuDNN custom call to have correct workspace size by build graph
-// and serialize so we can use it later
-class CuDnnWorkspaceRewriter : public HloModulePass {
+// Compile cuDNN custom calls to binaries and serialize them.
+// Also adjust them in HLO to have correct workspace size.
+class CuDnnCustomCallCompiler : public HloModulePass {
  public:
-  explicit CuDnnWorkspaceRewriter(se::StreamExecutor& stream_exec)
-      : dnn_support_(*stream_exec.AsDnn()) {}
+  explicit CuDnnCustomCallCompiler(se::StreamExecutor& stream_exec,
+                                   BinaryMap& compilation_results)
+      : dnn_support_(*stream_exec.AsDnn()),
+        compilation_results_(compilation_results) {}
 
-  absl::string_view name() const override { return "cudnn-workspace-rewriter"; }
+  absl::string_view name() const override {
+    return "cudnn-custom-call-compiler";
+  }
 
   using HloPassInterface::Run;
   absl::StatusOr<bool> Run(
@@ -43,9 +48,10 @@ class CuDnnWorkspaceRewriter : public HloModulePass {
 
  private:
   se::dnn::DnnSupport& dnn_support_;
+  BinaryMap& compilation_results_;
 };
 
 }  // namespace gpu
 }  // namespace xla
 
-#endif  // XLA_SERVICE_GPU_TRANSFORMS_CUDNN_WORKSPACE_REWRITER_H_
+#endif  // XLA_SERVICE_GPU_TRANSFORMS_CUDNN_CUSTOM_CALL_COMPILER_H_
