+++ xla/stream_executor/lazy_op_runner.h
@@ -280,76 +280,6 @@ struct FusedMatmulOp {
   }
 };
 
-struct FusedMHAOp {
-  using Signature = FusedMHASignature;
-  struct Config {
-    double scale;
-    const MatmulTensorDescriptor& bmm1_lhs_descriptor;
-    const MatmulTensorDescriptor& bmm1_rhs_descriptor;
-    const MatmulTensorDescriptor& bmm2_rhs_descriptor;
-    const MatmulTensorDescriptor& intermediate_bmm2_lhs_descriptor;
-    const TensorDescriptor& output_descriptor;
-    std::optional<TensorDescriptor> bias_descriptor;
-    std::optional<TensorDescriptor> activation_descriptor;
-    std::optional<double> dropout_rate;
-    std::optional<int64_t> seed;
-    FMHAMaskKind mask_type;
-  };
-
-  static absl::StatusOr<std::unique_ptr<const OpRunner<FusedMHASignature>>>
-  RunnerFromAlgorithmDesc(const AlgorithmDesc& desc, Config config,
-                          Stream* stream) {
-    TF_ASSIGN_OR_RETURN(auto dnn, internal::GetDnnFromStream(stream));
-    return dnn->FusedMHARunnerFromDesc(
-        stream, desc, config.bmm1_lhs_descriptor, config.bmm1_rhs_descriptor,
-        config.bmm2_rhs_descriptor, config.intermediate_bmm2_lhs_descriptor,
-        config.output_descriptor, config.activation_descriptor,
-        config.bias_descriptor, config.scale, config.dropout_rate, config.seed,
-        config.mask_type);
-  }
-};
-
-struct FusedMHABackwardOp {
-  using Signature = FusedMHABackwardSignature;
-
-  struct Config {
-    double scale;
-    const MatmulTensorDescriptor& bmm1_grad_gemm1_rhs_descriptor;
-    const MatmulTensorDescriptor& bmm1_grad_gemm2_rhs_descriptor;
-    const MatmulTensorDescriptor& bmm2_grad_gemm1_lhs_descriptor;
-    const MatmulTensorDescriptor& bmm2_grad_gemm2_rhs_descriptor;
-    const MatmulTensorDescriptor& d_output_descriptor;
-    const TensorDescriptor& d_bmm1_lhs_descriptor;
-    const TensorDescriptor& d_bmm1_rhs_descriptor;
-    const TensorDescriptor& d_bmm2_rhs_descriptor;
-    std::optional<TensorDescriptor> d_s_descriptor;
-    std::optional<TensorDescriptor> d_bias_descriptor;
-    std::optional<TensorDescriptor> fwd_output_descriptor;
-    std::optional<TensorDescriptor> bias_descriptor;
-    std::optional<double> dropout_rate;
-    std::optional<int64_t> seed;
-    FMHAMaskKind mask_type;
-    bool force_deterministic;
-  };
-
-  static absl::StatusOr<
-      std::unique_ptr<const OpRunner<FusedMHABackwardSignature>>>
-  RunnerFromAlgorithmDesc(const AlgorithmDesc& desc, Config config,
-                          Stream* stream) {
-    TF_ASSIGN_OR_RETURN(auto dnn, internal::GetDnnFromStream(stream));
-    return dnn->FusedMHABackwardRunnerFromDesc(
-        stream, desc, config.bmm1_grad_gemm1_rhs_descriptor,
-        config.bmm1_grad_gemm2_rhs_descriptor,
-        config.bmm2_grad_gemm1_lhs_descriptor,
-        config.bmm2_grad_gemm2_rhs_descriptor, config.d_output_descriptor,
-        config.d_bmm1_lhs_descriptor, config.d_bmm1_rhs_descriptor,
-        config.d_bmm2_rhs_descriptor, config.d_s_descriptor,
-        config.d_bias_descriptor, config.fwd_output_descriptor,
-        config.bias_descriptor, config.scale, config.dropout_rate, config.seed,
-        config.mask_type, config.force_deterministic);
-  }
-};
-
 }  // namespace dnn
 }  // namespace stream_executor
